clients/quickfix-client/pyproject.toml
[tool.poetry]
name = "broker-quickfix-client"
version = "0.3.0"
description = ""
authors = ["Samuel Guillemet <samuel.guillemet@telecom-sudparis.eu>"]
readme = "README.md"
packages = [{ include = "broker_quickfix_client" }]

[tool.poetry.dependencies]
python = "~3.11"
quickfix = "^1.15.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
mypy = "^1.4.1"
mypy-extensions = "^1.0.0"
pylint = "^2.17.4"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
types-pyyaml = "^6.0.12.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
no_strict_optional = true
ignore_missing_imports = true
files = "^(broker_quickfix_client/|tests/)"
